fixed yarn and download script

+++ helper/download.sh
@@ -10,7 +10,7 @@ download() {
     package_install curl
   fi
 
-  # if [ ! -f "$2" ]; then
+  if [ ! -f "$2" ]; then
     eval_cmd "$name" "curl -o $2 $1"
-  # fi
+  fi
 }
